Modularize fixture-matrix.mjs into composable step/collector modules
Pure behavior-preserving extraction (Refs #242, workstream 3). The accreted
fixture-matrix.mjs monolith is split into focused modules under
lib/fixture-matrix/ and fixture-matrix.mjs becomes a thin composer that
re-exports the same public surface from the same path. No logic changes —
functions and constants were relocated verbatim.

Modules:
- shared/constants.mjs, shared/utils.mjs (schemas, loss classes, helpers)
- fixtures.mjs (discovery + taxonomy)
- findings.mjs (#535 classification + honest loss-acceptance)
- steps/recipe-builder.mjs, steps/editor-validation-step.mjs (#537),
  steps/visual-parity-step.mjs (#538)
- collectors/editor-validation.mjs (#537), collectors/visual-parity.mjs (#538),
  collectors/quality-metrics.mjs (#541), collectors/run-intake.mjs
- result.mjs (quality gate + summary/aggregate rollups + artifact writing)

fixture-matrix.mjs: 2268 -> 57 lines.

Verification: tools/fixture-matrix.test.mjs 42 -> 42 pass; run-fixture-matrix.mjs
--dry-run composes the identical recipe/plan before and after (capture-only and
gated configs both byte-identical).
