+++ WordPress/static-site-importer/lib/fixture-matrix/collectors/run-intake.mjs
@@ -1,0 +1,245 @@
+// Run-result intake for the Static Site Importer fixture matrix: reads WP
+// Codebox runtime payloads + per-fixture artifact files back out, normalizes
+// them into fixture results, and threads the per-concern collectors together.
+//
+// Extracted verbatim from the former `lib/fixture-matrix.mjs` monolith as part
+// of the matrix modularization (Refs #242).
+
+import fs from 'node:fs';
+import path from 'node:path';
+
+import {
+  normalizeArray,
+  objectValue,
+  numberValue,
+  firstString,
+  compactObject,
+  mergeObjects,
+  diagnosticMessage,
+  requiredString,
+  readJsonFileIfExists,
+  artifactRef,
+  parseJsonPayloadsFromText,
+} from '../shared/utils.mjs';
+import { createFixtureMatrix } from '../fixtures.mjs';
+import { dedupeDiagnostics } from '../findings.mjs';
+import { collectQualityMetrics, collectBlockComposition } from './quality-metrics.mjs';
+import { collectEditorValidationDiagnostics } from './editor-validation.mjs';
+import {
+  collectVisualParityDiagnostics,
+  collectVisualParityArtifacts,
+  normalizeVisualParityGateOptions,
+} from './visual-parity.mjs';
+import { normalizeFixtureMatrixResult, normalizeFixtureResult } from '../result.mjs';
+
+export function collectFixtureMatrixRunResults(input = {}) {
+  const matrix = input.matrix || createFixtureMatrix(input);
+  const outputDirectory = requiredString(input.outputDirectory || input.output_directory, 'outputDirectory');
+  const codeboxOutput = input.codeboxOutput || input.codebox_output || readJsonFileIfExists(input.outputFile || input.output_file) || null;
+  const codeboxError = input.codeboxError || input.codebox_error || null;
+  const runtimePayloads = collectRuntimePayloads(codeboxOutput);
+  const visualParity = normalizeVisualParityGateOptions(input.visualParity || input.visual_parity || input);
+  const results = matrix.fixtures.map((fixture) => {
+    const fixtureArtifactsDirectory = path.join(outputDirectory, fixture.id);
+    const payloads = [
+      ...runtimePayloads.filter((payload) => fixtureIdentity(payload) === fixture.id),
+      ...readFixturePayloadFiles(fixtureArtifactsDirectory),
+    ];
+    return normalizeCollectedFixtureResult({ fixture, payloads, fixtureArtifactsDirectory, codeboxError, visualParity });
+  });
+
+  return normalizeFixtureMatrixResult({ matrix, results });
+}
+
+function normalizeCollectedFixtureResult({ fixture, payloads, fixtureArtifactsDirectory, codeboxError, visualParity }) {
+  const merged = mergeObjects(payloads);
+  const diagnostics = collectFixtureDiagnostics(merged, { visualParity });
+  const error = firstString([
+    merged.error,
+    merged.message && isFailurePayload(merged) ? merged.message : '',
+    codeboxError && payloads.length === 0 ? codeboxError.message || String(codeboxError) : '',
+  ]);
+  const success = inferFixtureSuccess(merged, diagnostics, error, payloads.length);
+  return normalizeFixtureResult({
+    fixture_id: fixture.id,
+    fixture_path: fixture.fixture_path,
+    status: fixtureStatus(payloads.length, error, success),
+    success,
+    error,
+    ssi_validation: merged.ssi_validation || merged.ssiValidation || merged.validation || merged.static_site_importer || null,
+    import_report: merged.import_report || merged.importReport || merged.report || null,
+    quality_metrics: collectQualityMetrics(merged),
+    block_composition: collectBlockComposition(merged),
+    blocks_engine_diagnostics: collectBlocksEngineDiagnostics(merged),
+    invalid_block_counts: collectInvalidBlockCounts(merged),
+    missing_assets: collectMissingAssets(merged),
+    runtime_target_gaps: collectRuntimeTargetGaps(merged),
+    diagnostics,
+    artifact_refs: collectFixtureArtifactRefs(merged, fixtureArtifactsDirectory),
+    artifacts: merged.artifacts || {},
+    visual_parity_artifacts: collectVisualParityArtifacts(merged),
+    raw: { payloads },
+  });
+}
+
+function collectFixtureDiagnostics(payload, options = {}) {
+  const diagnostics = [
+    ...normalizeArray(payload.diagnostics),
+    ...normalizeArray(payload.fixture_diagnostics?.diagnostics || payload.fixtureDiagnostics?.diagnostics),
+    ...normalizeArray(payload.findings),
+    ...collectFindingPacketDiagnostics(payload),
+    ...normalizeArray(payload.messages),
+    ...normalizeArray(payload.errors),
+    ...normalizeArray(payload.warnings),
+    ...normalizeArray(payload.upstream_gaps || payload.upstreamGaps).map((gap) => ({ kind: 'upstream_gap', ...objectValue(gap), message: diagnosticMessage(gap) || gap.missing || 'Upstream capability gap detected.' })),
+    ...collectBlocksEngineDiagnostics(payload),
+    ...collectRuntimeTargetGaps(payload).map((gap) => ({ kind: 'runtime_target_gap', ...objectValue(gap), message: diagnosticMessage(gap) || 'Runtime target gap detected.' })),
+    ...collectMissingAssets(payload).map((asset) => ({ kind: missingAssetKind(asset), ...objectValue(asset), message: diagnosticMessage(asset) || 'Missing imported asset.' })),
+    ...collectEditorValidationDiagnostics(payload),
+    ...collectVisualParityDiagnostics(payload, options.visualParity),
+  ];
+  const invalidBlockCount = Object.values(collectInvalidBlockCounts(payload)).reduce((sum, value) => sum + numberValue(value), 0);
+  if (invalidBlockCount > 0) {
+    diagnostics.push({ kind: 'invalid_block_content', message: `${invalidBlockCount} invalid block${invalidBlockCount === 1 ? '' : 's'} reported by SSI validation.` });
+  }
+  return dedupeDiagnostics(diagnostics);
+}
+
+function collectFindingPacketDiagnostics(payload) {
+  return [
+    ...normalizeArray(payload.finding_packets?.packets || payload.findingPackets?.packets),
+    ...normalizeArray(payload.import_report?.finding_packets?.packets || payload.importReport?.finding_packets?.packets),
+    ...normalizeArray(payload.report?.finding_packets?.packets),
+  ];
+}
+
+function collectFixtureArtifactRefs(payload, fixtureArtifactsDirectory) {
+  const refs = [...normalizeArray(payload.artifact_refs || payload.artifactRefs), ...normalizeArray(payload.artifacts?.refs)];
+  for (const [key, value] of Object.entries(payload.artifacts || {})) {
+    if (value && typeof value === 'object' && !Array.isArray(value) && (value.path || value.file || value.href)) {
+      refs.push({ artifact_id: key, kind: value.kind || key, ...value });
+    } else if (typeof value === 'string') {
+      refs.push({ artifact_id: key, kind: key, path: value });
+    }
+  }
+  for (const fileName of ['artifact.json', 'validation-result.json', 'import-report.json']) {
+    const filePath = path.join(fixtureArtifactsDirectory, fileName);
+    if (fs.existsSync(filePath)) {
+      refs.push(artifactRef(fileName.replace(/\.json$/, ''), filePath, fileName === 'artifact.json' ? 'input' : 'diagnostic'));
+    }
+  }
+  return refs;
+}
+
+function collectRuntimePayloads(value) {
+  const payloads = [];
+  visitRuntimePayloads(value, '', payloads, new Set());
+  return payloads;
+}
+
+function visitRuntimePayloads(value, inheritedFixtureId, payloads, seen) {
+  if (!value || typeof value !== 'object' || seen.has(value)) {
+    return;
+  }
+  seen.add(value);
+  const fixtureId = fixtureIdentity(value) || inheritedFixtureId;
+  if (fixtureId && hasPayloadData(value)) {
+    payloads.push({ fixture_id: fixtureId, ...value });
+  }
+  for (const key of ['stdout', 'stderr', 'output', 'result']) {
+    for (const parsed of parseJsonPayloadsFromText(value[key])) {
+      payloads.push({ fixture_id: fixtureId, ...parsed });
+    }
+  }
+  for (const child of Array.isArray(value) ? value : Object.values(value)) {
+    visitRuntimePayloads(child, fixtureId, payloads, seen);
+  }
+}
+
+function hasPayloadData(value) {
+  return ['status', 'success', 'ok', 'passed', 'error', 'diagnostics', 'findings', 'summary', 'artifacts', 'upstream_gaps', 'runtime_target_gaps', 'blocks_engine', 'import_report']
+    .some((key) => Object.hasOwn(value, key));
+}
+
+function readFixturePayloadFiles(directory) {
+  return ['validation-result.json', 'result.json', 'import-report.json', 'quality.json', 'blocks-engine-diagnostics.json', 'editor-validation.json', 'editor-canvas-summary.json', 'visual-compare.json', 'visual-diff.json', 'visual-parity.json']
+    .map((fileName) => readJsonFileIfExists(path.join(directory, fileName)))
+    .filter(Boolean);
+}
+
+function fixtureIdentity(payload) {
+  return payload?.fixture_id
+    || payload?.fixtureId
+    || payload?.fixture?.id
+    || payload?.fixture?.slug
+    || payload?.fixture_diagnostics?.fixture?.slug
+    || payload?.fixtureDiagnostics?.fixture?.slug
+    || payload?.request?.import_args?.slug
+    || payload?.request?.importArgs?.slug
+    || payload?.metadata?.fixture_id
+    || payload?.metadata?.fixtureId
+    || '';
+}
+
+function collectInvalidBlockCounts(payload) {
+  const quality = collectQualityMetrics(payload);
+  return compactObject({
+    invalid_block_count: payload.invalid_block_count || payload.invalidBlockCount || quality.invalid_block_count,
+    invalid_blocks: payload.invalid_blocks || payload.invalidBlocks || quality.invalid_blocks,
+    editor_invalid_blocks: payload.editor_invalid_blocks || payload.editorInvalidBlocks || quality.editor_invalid_blocks,
+  });
+}
+
+function collectMissingAssets(payload) {
+  return [
+    ...normalizeArray(payload.missing_assets || payload.missingAssets),
+    ...normalizeArray(payload.dropped_images || payload.droppedImages),
+    ...normalizeArray(payload.import_report?.missing_assets || payload.importReport?.missing_assets),
+    ...normalizeArray(payload.report?.missing_assets),
+  ];
+}
+
+function collectRuntimeTargetGaps(payload) {
+  return [
+    ...normalizeArray(payload.runtime_target_gaps || payload.runtimeTargetGaps),
+    ...normalizeArray(payload.runtime_targets_missing || payload.runtimeTargetsMissing),
+    ...normalizeArray(payload.blocks_engine?.runtime_target_gaps || payload.blocksEngine?.runtimeTargetGaps),
+  ];
+}
+
+function collectBlocksEngineDiagnostics(payload) {
+  return [
+    ...normalizeArray(payload.blocks_engine_diagnostics || payload.blocksEngineDiagnostics),
+    ...normalizeArray(payload.blocks_engine?.diagnostics || payload.blocksEngine?.diagnostics),
+    ...normalizeArray(payload.transformer_diagnostics || payload.transformerDiagnostics),
+  ];
+}
+
+function inferFixtureSuccess(payload, diagnostics, error, payloadCount) {
+  if (payload.success === true || payload.ok === true || payload.passed === true) {
+    return diagnostics.length === 0 && !error;
+  }
+  if (payload.success === false || payload.ok === false || payload.passed === false || payload.status === 'failed' || payload.status === 'error') {
+    return false;
+  }
+  if (payload.status === 'passed' || payload.status === 'success') {
+    return diagnostics.length === 0 && !error;
+  }
+  return payloadCount > 0 && diagnostics.length === 0 && !error;
+}
+
+function fixtureStatus(payloadCount, error, success) {
+  if (payloadCount === 0 && !error) {
+    return 'not_run';
+  }
+  return success ? 'passed' : 'failed';
+}
+
+function isFailurePayload(payload) {
+  return payload.success === false || payload.ok === false || payload.status === 'failed' || payload.status === 'error';
+}
+
+function missingAssetKind(value) {
+  const message = diagnosticMessage(value);
+  return /\.svg(?:\b|$)/i.test(message) ? 'broken_svg' : 'dropped_images';
+}
