+++ WordPress/static-site-importer/lib/fixture-matrix/steps/recipe-builder.mjs
@@ -1,0 +1,133 @@
+// WP Codebox recipe building (import + editor-validation + visual-parity steps)
+// and fixture-artifact construction for the Static Site Importer fixture matrix.
+//
+// Extracted verbatim from the former `lib/fixture-matrix.mjs` monolith as part
+// of the matrix modularization (Refs #242).
+
+import fs from 'node:fs';
+import path from 'node:path';
+
+import {
+  WEBSITE_ARTIFACT_SCHEMA,
+  DEFAULT_ENTRYPOINT,
+  DEFAULT_IMPORTER_SLUG,
+  DEFAULT_EDITOR_VALIDATION_URL,
+} from '../shared/constants.mjs';
+import {
+  normalizeArray,
+  isTextPayloadType,
+  isImagePath,
+  requiredString,
+  shellToken,
+} from '../shared/utils.mjs';
+import { createFixtureMatrix, normalizeFixture, collectFixtureFiles } from '../fixtures.mjs';
+import { editorBlockValidationStep } from './editor-validation-step.mjs';
+import { visualParityCompareStep, normalizeVisualParityRecipeOptions } from './visual-parity-step.mjs';
+
+export function buildFixtureArtifact(fixture, options = {}) {
+  const normalized = normalizeFixture(fixture);
+  const files = collectFixtureFiles(normalized.directory, options);
+  const artifactFiles = files.map((file) => {
+    const payload = fs.readFileSync(file.absolute_path);
+    const artifactFile = {
+      path: `website/${file.relative_path}`,
+      source_path: file.absolute_path,
+      type: file.type,
+      bytes: file.bytes,
+    };
+    if (isTextPayloadType(file.type)) {
+      artifactFile.content = payload.toString('utf8');
+    } else {
+      artifactFile.content_base64 = payload.toString('base64');
+    }
+    return artifactFile;
+  });
+
+  return {
+    schema: WEBSITE_ARTIFACT_SCHEMA,
+    entrypoint: DEFAULT_ENTRYPOINT,
+    entry_path: DEFAULT_ENTRYPOINT,
+    files: artifactFiles,
+    summary: {
+      file_count: artifactFiles.length,
+      entry_path: DEFAULT_ENTRYPOINT,
+      has_css: artifactFiles.some((file) => file.path.endsWith('.css')),
+      has_js: artifactFiles.some((file) => file.path.endsWith('.js')),
+      has_images: artifactFiles.some((file) => isImagePath(file.path)),
+    },
+    source_metadata: {
+      fixture_id: normalized.id,
+      fixture_path: normalized.directory,
+      fixture_entrypoint: normalized.entrypoint,
+    },
+  };
+}
+
+export function buildFixtureMatrixRecipe(input = {}) {
+  const matrix = input.matrix || createFixtureMatrix(input);
+  const artifactsDirectory = input.artifactsDirectory || input.artifacts_directory || '/artifacts/static-site-importer-fixture-matrix';
+  const playgroundArtifactsDirectory = input.playgroundArtifactsDirectory || input.playground_artifacts_directory;
+  const commandArtifactsDirectory = playgroundArtifactsDirectory || artifactsDirectory;
+  const importer = normalizeStaticSiteImporterPlugin(input);
+  const mounts = normalizeArray(input.mounts);
+  const extraPlugins = [importer.extraPlugin, ...normalizeArray(input.extraPlugins || input.extra_plugins)];
+  const editorValidationEnabled = input.editorValidation !== false && input.editor_validation !== false;
+  const editorValidationUrl = input.editorValidationUrl || input.editor_validation_url || DEFAULT_EDITOR_VALIDATION_URL;
+  const visualParityEnabled = input.visualParity !== false && input.visual_parity !== false;
+  const visualParityRecipeOptions = normalizeVisualParityRecipeOptions(input);
+
+  if (playgroundArtifactsDirectory) {
+    mounts.push({
+      source: artifactsDirectory,
+      target: playgroundArtifactsDirectory,
+      mode: 'readwrite',
+    });
+  }
+
+  return {
+    schema: 'wp-codebox/workspace-recipe/v1',
+    runtime: {
+      wp: input.wordpressVersion || input.wordpress_version || 'latest',
+      blueprint: input.blueprint || {},
+    },
+    inputs: {
+      mounts,
+      extra_plugins: extraPlugins,
+    },
+    workflow: {
+      steps: [
+        importer.activationStep,
+        ...matrix.fixtures.flatMap((fixture) => [
+          {
+            command: 'wordpress.wp-cli',
+            args: [
+              `command=static-site-importer validate-artifact --artifact=${shellToken(path.join(commandArtifactsDirectory, fixture.id, 'artifact.json'))} --slug=${shellToken(fixture.id)} --name=${shellToken(fixture.label)} --allow-missing-woocommerce --allow-failure`,
+            ],
+          },
+          ...(editorValidationEnabled ? [editorBlockValidationStep({ fixture, url: editorValidationUrl })] : []),
+          ...(visualParityEnabled ? [visualParityCompareStep({ fixture, ...visualParityRecipeOptions })] : []),
+        ]),
+      ],
+    },
+    artifacts: {
+      directory: artifactsDirectory,
+    },
+  };
+}
+
+export function normalizeStaticSiteImporterPlugin(input = {}) {
+  const source = requiredString(input.staticSiteImporterPath || input.static_site_importer_path, 'staticSiteImporterPath');
+  const slugValue = input.staticSiteImporterSlug || input.static_site_importer_slug || DEFAULT_IMPORTER_SLUG;
+  const pluginFile = input.staticSiteImporterPlugin || input.static_site_importer_plugin || `${slugValue}/${slugValue}.php`;
+  return {
+    extraPlugin: {
+      source,
+      slug: slugValue,
+      activate: true,
+    },
+    activationStep: {
+      command: 'wordpress.wp-cli',
+      args: [`command=plugin activate ${pluginFile}`],
+    },
+  };
+}
